+++ src/MBH/Bundle/ChannelManagerBundle/Services/Ostrovok.php
@@ -13,7 +13,6 @@ use MBH\Bundle\PackageBundle\Document\Order;
 use MBH\Bundle\PackageBundle\Document\Package;
 use MBH\Bundle\PackageBundle\Document\PackagePrice;
 use MBH\Bundle\PackageBundle\Document\PackageService;
-use MBH\Bundle\PriceBundle\Document\PriceCache;
 use MBH\Bundle\PriceBundle\Document\Restriction;
 use MBH\Bundle\PriceBundle\Document\RoomCache;
 use Symfony\Component\DependencyInjection\ContainerInterface;
@@ -188,7 +187,7 @@ class Ostrovok extends Base
             }
         }
 
-        return $this->sendApiRequest($rna_request_data, __METHOD__);
+        return $this->sendApiRequest($rna_request_data, __METHOD__) && $result;
     }
 
     /**
@@ -263,6 +262,7 @@ class Ostrovok extends Base
                                         '%numberOfAdults%' => $adults
                                     ]);
                                     $this->log('Error! '.$message);
+                                    $result = false;
                                     continue;
                                     /*throw new \Exception($message);*/
                                 }
@@ -304,7 +304,7 @@ class Ostrovok extends Base
             }
         }
 
-        return $this->sendApiRequest($rna_request_data, __METHOD__);
+        return $this->sendApiRequest($rna_request_data, __METHOD__) && $result;
     }
 
     /**
@@ -341,13 +341,13 @@ class Ostrovok extends Base
 
 
             $configRoomTypes = $this->getRoomTypes($config);
-            $configTarrifs = $this->getTariffs($config);
+            $configTariffs = $this->getTariffs($config);
             $roomTypeRestrictions = $this->dm->getRepository('MBHPriceBundle:Restriction')->fetch(
                 $begin,
                 $end,
                 $config->getHotel(),
                 $roomType ? [$roomType->getId()] : array_keys($configRoomTypes),
-                array_keys($configTarrifs),
+                array_keys($configTariffs),
                 true
             );
 
@@ -417,7 +417,7 @@ class Ostrovok extends Base
             }
         }
 
-        return $this->sendApiRequest($rna_request_data, __METHOD__);
+        return $this->sendApiRequest($rna_request_data, __METHOD__) && $result;
     }
 
     /**