+++ src/MBH/Bundle/ClientBundle/Document/ClientConfig.php
@@ -3,16 +3,13 @@
 namespace MBH\Bundle\ClientBundle\Document;
 
 use Doctrine\ODM\MongoDB\Mapping\Annotations as ODM;
-use Doctrine\ODM\MongoDB\PersistentCollection;
 use Gedmo\Mapping\Annotation as Gedmo;
 use Gedmo\SoftDeleteable\Traits\SoftDeleteableDocument;
 use Gedmo\Timestampable\Traits\TimestampableDocument;
 use MBH\Bundle\BaseBundle\Document\Base;
-use MBH\Bundle\BaseBundle\Document\NotificationType;
 use MBH\Bundle\BaseBundle\Document\Traits\AllowNotificationTypesTrait;
 use MBH\Bundle\BaseBundle\Document\Traits\BlameableDocument;
 use MBH\Bundle\CashBundle\Document\CashDocument;
-use MBH\Bundle\ClientBundle\Traits\CheckIsNotificationTypeExistsTrait;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Validator\Constraints as Assert;
 
@@ -24,6 +21,7 @@ use Symfony\Component\Validator\Constraints as Assert;
 class ClientConfig extends Base
 {
     const DEFAULT_NUMBER_OF_DAYS_FOR_PAYMENT = 5;
+    const DEFAULT_BEGIN_DATE_OFFSET = -21;
 
     /**
      * Hook timestampable behavior
@@ -182,6 +180,13 @@ class ClientConfig extends Base
     protected $beginDate;
 
     /**
+     * @var int
+     * @ODM\Field(type="int")
+     * @Assert\Type(type="int")
+     */
+    protected $beginDateOffset;
+
+    /**
      * @var integer
      * @Gedmo\Versioned
      * @ODM\Integer()
@@ -190,6 +195,25 @@ class ClientConfig extends Base
      * @Assert\Range(min=0, max=365)
      */
     protected $noticeUnpaid = 0;
+
+    /**
+     * @var integer
+     * @Gedmo\Versioned
+     * @ODM\Field(type="integer")
+     * @Assert\NotNull()
+     * @Assert\Type(type="integer", message="validate.type.integer")
+     * @Assert\Range(min=0, max=2, invalidMessage="")
+     */
+    protected $priceRoundSign = 2;
+
+    /**
+     * @var boolean
+     * @Gedmo\Versioned
+     * @ODM\Boolean()
+     * @Assert\NotNull()
+     * @Assert\Type(type="boolean")
+     */
+    protected $queryStat = true;
 
     /**
      * @var bool
@@ -694,6 +718,75 @@ class ClientConfig extends Base
     public function setBeginDate(?\DateTime $beginDate)
     {
         $this->beginDate = $beginDate;
+    }
+
+    /**
+     * @return integer
+     */
+    public function getPriceRoundSign()
+    {
+        return $this->priceRoundSign;
+    }
+
+    /**
+     * @param integer $priceRoundSign
+     */
+    public function setPriceRoundSign($priceRoundSign)
+    {
+        $this->priceRoundSign = $priceRoundSign;
+    }
+
+
+    /**
+     * @return int
+     */
+    public function getBeginDateOffset(): ?int
+    {
+        return $this->beginDateOffset;
+    }
+
+    /**
+     * @param int $beginDateOffset
+     * @return ClientConfig
+     */
+    public function setBeginDateOffset(?int $beginDateOffset): ClientConfig
+    {
+        $this->beginDateOffset = $beginDateOffset;
+
+        return $this;
+    }
+
+    /**
+     * @return \DateTime
+     */
+    public function getActualBeginDate()
+    {
+        if (!empty($this->getBeginDate())) {
+            return $this->getBeginDate();
+        }
+
+        $beginDateOffset = !empty($this->getBeginDateOffset()) ? $this->getBeginDateOffset() : self::DEFAULT_BEGIN_DATE_OFFSET;
+
+        return (new \DateTime('midnight'))->modify($beginDateOffset . ' days');
+    }
+
+    /**
+     * @return bool
+     */
+    public function isQueryStat(): bool
+    {
+        return $this->queryStat;
+    }
+
+    /**
+     * @param bool $queryStat
+     * @return ClientConfig
+     */
+    public function setQueryStat(bool $queryStat): ClientConfig
+    {
+        $this->queryStat = $queryStat;
+
+        return $this;
     }
 
     /**