+++ src/MBH/Bundle/HotelBundle/Document/Hotel.php
@@ -402,6 +402,36 @@ class Hotel extends Base implements \JsonSerializable, AddressInterface
      */
     protected $packageDepartureTime = self::DEFAULT_DEPARTURE_TIME;
 
+    /**
+     * @var string
+     * @ODM\Field(type="string")
+     */
+    protected $aboutLink;
+
+    /**
+     * @var string
+     * @ODM\Field(type="string")
+     */
+    protected $roomsLink;
+
+    /**
+     * @var string
+     * @ODM\Field(type="string")
+     */
+    protected $mapLink;
+
+    /**
+     * @var string
+     * @ODM\Field(type="string")
+     */
+    protected $contactsLink;
+
+    /**
+     * @var string
+     * @ODM\Field(type="string")
+     */
+    protected $pollLink;
+
     public function __construct()
     {
         $this->roomTypes = new ArrayCollection();
@@ -1665,6 +1695,100 @@ class Hotel extends Base implements \JsonSerializable, AddressInterface
         return $this;
     }
 
+    /**
+     * @return string
+     */
+    public function getAboutLink(): ?string
+    {
+        return $this->aboutLink;
+    }
+
+    /**
+     * @param string $aboutLink
+     * @return Hotel
+     */
+    public function setAboutLink(string $aboutLink): Hotel
+    {
+        $this->aboutLink = $aboutLink;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getRoomsLink(): ?string
+    {
+        return $this->roomsLink;
+    }
+
+    /**
+     * @param string $roomsLink
+     * @return Hotel
+     */
+    public function setRoomsLink(string $roomsLink): Hotel
+    {
+        $this->roomsLink = $roomsLink;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getMapLink(): ?string
+    {
+        return $this->mapLink;
+    }
+
+    /**
+     * @param string $mapLink
+     * @return Hotel
+     */
+    public function setMapLink(string $mapLink): Hotel
+    {
+        $this->mapLink = $mapLink;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getContactsLink(): ?string
+    {
+        return $this->contactsLink;
+    }
+
+    /**
+     * @param string $contactsLink
+     * @return Hotel
+     */
+    public function setContactsLink(string $contactsLink): Hotel
+    {
+        $this->contactsLink = $contactsLink;
+
+        return $this;
+    }
+
+    /**
+     * @return string
+     */
+    public function getPollLink(): ?string
+    {
+        return $this->pollLink;
+    }
+
+    /**
+     * @param string $pollLink
+     * @return Hotel
+     */
+    public function setPollLink(string $pollLink): Hotel
+    {
+        $this->pollLink = $pollLink;
+
+        return $this;
+    }
     /**
      * @return int
      */